backend/requirements.txt
fastapi==0.115.12
uvicorn==0.24.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.0.76
timm==0.6.13
python-multipart==0.0.6
typing-extensions>=4.8.0
starlette>=0.27.0
pydantic>=1.7.4,<3.0.0
